+++ CHANGELOG.md
@@ -6,6 +6,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Unreleased
 
+### Features
+
+- Added TLV parser for `IBILteCellInfoT`
+
 ## [v2.0.0](https://github.com/seemoo-lab/aristoteles/tree/v2.0.0) (28-11-2021)
 
 ### Features